move param matching to helper function
